Move function definitions to source code files
