+++ src/General.hpp
@@ -1,0 +1,10 @@
+//
+// Created by hubert25632 on 01.07.22.
+//
+
+#ifndef BATTLESHIPS_GENERAL_HPP
+#define BATTLESHIPS_GENERAL_HPP
+
+#define XOFFSET 40
+
+#endif //BATTLESHIPS_GENERAL_HPP
