+++ src/Timer.hpp
@@ -5,6 +5,8 @@
 #ifndef BATTLESHIPS_TIMER_HPP
 #define BATTLESHIPS_TIMER_HPP
 
+#include <pthread.h>
+
 class Timer {
 private:
     pthread_t timer;
@@ -13,47 +15,16 @@ private:
     bool threadOk = true;
     bool barrierOk = true;
 
-    static void *timerCallback(void *barrier) {
-        pthread_barrier_t *barrier_c = (pthread_barrier_t *) barrier;
-        while (true) {
-            pthread_barrier_wait(barrier_c);
-            SDL_Delay(100);
-        }
-    }
+    static void *timerCallback(void *barrier);
 
 public:
-    Timer() {
-        int e;
-        e = pthread_barrier_init(&barrier, NULL, 2);
-        if (e != 0) {
-            errorMessage(std::string("pthread_barrier_init: ") + std::string(strerror(e)));
-            timerOk = false;
-            barrierOk = false;
-            return;
-        }
-        e = pthread_create(&timer, NULL, timerCallback, &barrier);
-        if (e != 0) {
-            errorMessage(std::string("pthread_create: ") + std::string(strerror(e)));
-            timerOk = false;
-            threadOk = false;
-            return;
-        }
-    }
+    Timer();
 
-    ~Timer() {
-        if (threadOk)
-            pthread_cancel(timer);
-        if (barrierOk)
-            pthread_barrier_destroy(&barrier);
-    }
+    ~Timer();
 
-    bool ok() {
-        return timerOk;
-    }
+    bool ok();
 
-    void synchronize() {
-        pthread_barrier_wait(&barrier);
-    }
+    void synchronize();
 };
 
 #endif //BATTLESHIPS_TIMER_HPP
